♻️ refactor(Wallet): Improve wallet balance handling.
The `balance` attribute is now handled as an integer and uses a new `raw_balance` attribute for internal calculations.

+++ src/Models/WalletTransaction.php
@@ -20,7 +20,7 @@ class WalletTransaction extends Model
 
     protected $casts = [
         'metadata' => 'array',
-        'amount' => 'float',
-        'running_balance' => 'float',
+        'amount' => 'int',
+        'running_balance' => 'int',
     ];
 }
